pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "weekly_review_report_adk"
version = "0.1.0"
description = "Hand-coded ADK 2.0 style framework MVP"
requires-python = ">=3.11"


dependencies = [
    "google-adk>=2.0.0a1",
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0",
  "ruff>=0.8.0",
  "mypy>=1.13",
]

monitoring = [
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-gcp-trace>=1.8",
]

llm = [
  "google-genai>=1.20.0",
]

[project.scripts]
adk2 = "adk2.cli.main:cli"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["src/weekly_project_report"]
